fix : fix ORM query and delete .gitmessage.txt
