--- .gitmessage.txt
@@ -1,26 +1,0 @@
-################
-# <타입> : <제목> 의 형식으로 제목을 아래 공백줄에 작성 (필수)
-# 제목은 50자 이내로 아랫줄에 영문 명령문으로 작성(첫글자 대문자) / 변경사항이 "무엇"인지 명확하게 작성 / 끝에 마침표 금지
-# (예시) feat : login
-
-# 바로 아래 공백은 지우지 마세요 (제목과 본문의 분리를 위함)
-
-################
-# 본문(구체적인 내용)을 아랫줄에 작성 (선택사항)
-# 여러 줄의 메시지를 작성할 땐 "-"로 구분 (한 줄은 72자 이내)
-
-################
-# 꼬릿말(footer)을 아랫줄에 작성 (선택사항) (현재 commit과 관련된 이슈 번호 등 추가)
-# (예시) Close #5
-
-################
-# <타입 목록>
-# feat : 새로운 기능 추가
-# fix : 버그 수정
-# docs : 문서화 및 문서 수정
-# style : 코드 의미에 영향을 주지 않는 변경사항 (코드 포맷팅, 세미콜론 누락 등)
-# refactor : 코드 리팩토링
-# test : 테스트 코드 추가 및 수정
-# chore : 빌드 부분 혹은 패키지 매니저 수정사항 (유지보수 및 그 외 수정사항)
-# build : 빌드나, 외부 dependency에 영향 준 변경사항 (ex. npm, gulp 등)
-################
